revived from git clean fdx
