+++ app/javascript/packs/application.js
@@ -3,16 +3,19 @@
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
 
+import Rails from '@rails/ujs';
+Rails.start();
 
 import 'bootstrap';
 import '../stylesheets/application.scss';
 import './bootstrap_custom.js'
 
-require("bootstrap");
-require("@rails/ujs").start()
+
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
+require('jquery')
+require("bootstrap");
 require('load_animation.js')
