+++ app/javascript/packs/rails-ujs.js
@@ -1,0 +1,2 @@
+import Rails from 'rails-ujs';
+Rails.start();
