+++ config/webpack/environment.js
@@ -1,18 +1,12 @@
 const { environment } = require('@rails/webpacker')
-const { VueLoaderPlugin } = require('vue-loader')
-const vue = require('./loaders/vue')
-
-environment.plugins.prepend('VueLoaderPlugin', new VueLoaderPlugin())
-environment.loaders.prepend('vue', vue)
-module.exports = environment
-
 
 const webpack = require('webpack')
-environment.plugins.append(
-  'Provide',
-  new webpack.ProvidePlugin({
-    $: 'jquery',
-    jQuery: 'jquery',
-    Popper: ['popper.js', 'default']
-  })
-)
+environment.plugins.prepend('Provide', new webpack.ProvidePlugin({
+  $: 'jquery/src/jquery',
+  jQuery: 'jquery/src/jquery',
+  jquery: 'jquery',
+  'window.jQuery': 'jquery',
+  Popper: ['popper.js', 'default']
+}))
+
+module.exports = environment
