+++ package.json
@@ -4,19 +4,20 @@
   "dependencies": {
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
-    "@rails/ujs": "^6.0.0",
+    "@rails/ujs": "^6.0.3-3",
     "@rails/webpacker": "4.2.2",
     "bootstrap": "4.3.1",
     "jquery": "^3.5.1",
     "popper.js": "^1.16.1",
+    "rails-ujs": "^5.2.4-4",
     "turbolinks": "^5.2.0",
-    "vue": "^2.6.12",
     "vue-loader": "^15.9.3",
     "vue-template-compiler": "^2.6.12",
     "vuetify": "^2.3.10"
   },
   "version": "0.1.0",
   "devDependencies": {
-    "webpack-dev-server": "^3.11.0"
+    "webpack-dev-server": "^3.11.0",
+    "webpack-manifest-plugin": "^2.2.0"
   }
 }
