--- spec/requests/access_to_sessions_spec.rb
@@ -1,56 +1,0 @@
-require 'rails_helper'
-require 'pp'
-require 'pry'
-
-RSpec.describe 'access to sessions', type: :request do
-  let(:customized_user) { FactoryBot.create(:customized_user) }
-  let(:other_user) { FactoryBot.create(:other_user) }
-
-  # ログインに成功
-  it "user successfully login" do
-    sign_in_as customized_user
-    expect(response).to redirect_to user_path(customized_user)
-    expect(page).to have_text("他のユーザー")
-  end
-
-
-  before do
-      visit login_path
-    end
-
-    describe 'enter an valid values' do
-      let!(:user) { create(:user, email: 'loginuser@example.com', password: 'password') }
-      before do
-        fill_in 'Email', with: 'loginuser@example.com'
-        fill_in 'Password', with: 'password'
-        click_button 'Log in'
-      end
-
-
-  # 無効な情報ではログインに失敗
-  it "user doesn't login with invalid information" do
-    visit login_path
-    fill_in "メールアドレス", with: ""
-    fill_in "パスワード", with: ""
-    click_button "ログインする！"
-    expect(current_path).to eq login_path
-    page.has_content? ("メールアドレスかパスワードが正しくありません。")
-  end
-
-describe 'DELETE #destroy' do
-  it 'log out and redirect to root page' do
-    delete logout_path
-    expect(response).to redirect_to root_path
-    #セッションの値のテスト
-    expect(is_logged_in?).to be_falsey
-
-    #もう一回ログアウトしてみる！
-    delete logout_path
-    follow_redirect!
-
-    expect(response).to have_http_status(200)
-  end
-end
-
-end
-end
